Change `origin_main`'s signature to avoid `c_int`.
Change `argc` to `usize` because a common use case is to pass it to
`slice::from_raw_parts`.

Change the return value to `i32`. This matches the type used by
`std::process::exit`.

+++ example-crates/external-start/src/main.rs
@@ -56,7 +56,7 @@ static EARLY_INIT_ARRAY: unsafe extern "C" fn(i32, *mut *mut u8) = {
 };
 
 #[no_mangle]
-unsafe fn origin_main(_argc: i32, _argv: *mut *mut u8, _envp: *mut *mut u8) -> i32 {
+unsafe fn origin_main(_argc: usize, _argv: *mut *mut u8, _envp: *mut *mut u8) -> i32 {
     eprintln!("Hello from main thread");
 
     at_exit(Box::new(|| eprintln!("Hello from an at_exit handler")));
